requirements.txt
alembic==1.8.1
Authlib==1.1.0
Babel==2.12.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.2
flask-babel==3.1.0
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
pytailwindcss==0.2.0
python-dotenv==0.21.0
pytz==2023.3.post1
requests==2.28.1
segno==1.5.2
SQLAlchemy==1.4.41
urllib3==1.26.12
Werkzeug==2.2.3
